requirements.txt
numpy
pyrtlsdr
